[feat] 툴바 프로필 및 검색창 레이아웃 구현
