+++ app/src/main/java/com/hyunjung/twitterclone/activities/LoginActivity.kt
@@ -1,4 +1,4 @@
-package com.hyunjung.twitterclone
+package com.hyunjung.twitterclone.activities
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -12,15 +12,11 @@ import android.widget.EditText
 import android.widget.Toast
 import com.google.android.material.textfield.TextInputLayout
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.firestore.FirebaseFirestore
-import com.hyunjung.twitterclone.databinding.ActivitySignupBinding
-import com.hyunjung.twitterclone.util.DATA_USERS
-import com.hyunjung.twitterclone.util.User
+import com.hyunjung.twitterclone.databinding.ActivityLoginBinding
 
-class SignupActivity : AppCompatActivity() {
+class LoginActivity : AppCompatActivity() {
 
-    private lateinit var binding: ActivitySignupBinding
-    private val firebaseDB = FirebaseFirestore.getInstance()
+    private lateinit var binding: ActivityLoginBinding
     private val firebaseAuth = FirebaseAuth.getInstance()
     private val firebaseAuthListener = FirebaseAuth.AuthStateListener {
         val user = firebaseAuth.currentUser?.uid
@@ -33,15 +29,14 @@ class SignupActivity : AppCompatActivity() {
     @SuppressLint("ClickableViewAccessibility")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivitySignupBinding.inflate(layoutInflater)
+        binding = ActivityLoginBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        setTextChangeListener(binding.usernameET, binding.usernameTIL)
         setTextChangeListener(binding.emailET, binding.emailTIL)
         setTextChangeListener(binding.passwordET, binding.passwordTIL)
 
         // ProgressBar가 떠있는 동안 touch 안되게 막아놓는 역할
-        binding.signupProgressLayout.setOnTouchListener{ v, event -> true }
+        binding.loginProgressLayout.setOnTouchListener{ v, event -> true }
     }
 
     private fun setTextChangeListener(et: EditText, til: TextInputLayout) {
@@ -56,13 +51,8 @@ class SignupActivity : AppCompatActivity() {
         })
     }
 
-    fun onSignup(v: View) {
+    fun onLogin(v: View) {
         var proceed = true
-        if (binding.usernameET.text.isNullOrBlank()) {
-            binding.usernameTIL.error = "Username is required"
-            binding.usernameTIL.isErrorEnabled = true
-            proceed = false
-        }
         if (binding.emailET.text.isNullOrBlank()) {
             binding.emailTIL.error = "Email is required"
             binding.emailTIL.isErrorEnabled = true
@@ -74,29 +64,23 @@ class SignupActivity : AppCompatActivity() {
             proceed = false
         }
         if (proceed) {
-            binding.signupProgressLayout.visibility = View.VISIBLE
-            firebaseAuth.createUserWithEmailAndPassword(binding.emailET.text.toString(), binding.passwordET.text.toString())
+            binding.loginProgressLayout.visibility = View.VISIBLE
+            firebaseAuth.signInWithEmailAndPassword(binding.emailET.text.toString(), binding.passwordET.text.toString())
                 .addOnCompleteListener { task ->
                     if (!task.isSuccessful) {
-                        Toast.makeText(this, "Signup error: ${task.exception?.localizedMessage}", Toast.LENGTH_SHORT).show()
-                    } else {
-                        val email = binding.emailET.text.toString()
-                        val name = binding.usernameET.text.toString()
-                        val user = User(email, name, "", arrayListOf(), arrayListOf())
-                        firebaseDB.collection(DATA_USERS).document(firebaseAuth.uid!!).set(user) // user 파이어베이스 DB 생성
+                        binding.loginProgressLayout.visibility = View.GONE
+                        Toast.makeText(this, "Login error: ${task.exception?.localizedMessage}", Toast.LENGTH_SHORT).show()
                     }
-                    binding.signupProgressLayout.visibility = View.GONE
                 }
                 .addOnFailureListener { e ->
                     e.printStackTrace()
-                    binding.signupProgressLayout.visibility = View.GONE
+                    binding.loginProgressLayout.visibility = View.GONE
                 }
         }
-
     }
 
-    fun goToLogin(v: View) {
-        startActivity(LoginActivity.newIntent(this))
+    fun goToSignup(v: View) {
+        startActivity(SignupActivity.newIntent(this))
         finish()
     }
 
@@ -111,6 +95,6 @@ class SignupActivity : AppCompatActivity() {
     }
 
     companion object {
-        fun newIntent(context: Context) = Intent(context, SignupActivity::class.java)
+        fun newIntent(context: Context) = Intent(context, LoginActivity::class.java)
     }
 }
