+++ app/src/main/res/layout/activity_home.xml
@@ -4,15 +4,77 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeActivity">
+    tools:context=".activities.HomeActivity">
+
+    <androidx.cardview.widget.CardView
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:background="@color/twitter_blue"
+        android:orientation="horizontal"
+        app:cardElevation="4dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
+
+        <androidx.cardview.widget.CardView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/grid_1x"
+            app:cardCornerRadius="@dimen/standard_corner_radius">
+
+            <ImageView
+                android:id="@+id/logo"
+                android:layout_width="@dimen/standard_image"
+                android:layout_height="@dimen/standard_image"
+                android:scaleType="centerCrop"
+                android:src="@drawable/default_user" />
+
+        </androidx.cardview.widget.CardView>
+
+        <TextView
+            android:id="@+id/titleBar"
+            style="@style/Heading"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center_vertical"
+            android:layout_marginStart="@dimen/standard_image"
+            android:paddingStart="@dimen/grid_3x"
+            android:text="Home" />
+
+        <androidx.cardview.widget.CardView
+            android:id="@+id/searchBar"
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:layout_gravity="center_vertical"
+            app:cardCornerRadius="@dimen/standard_corner_radius"
+            android:layout_marginLeft="60dp"
+            android:layout_marginStart="60dp">
+
+            <EditText
+                android:id="@+id/search"
+                android:layout_width="@dimen/standard_width"
+                android:layout_height="@dimen/search_height"
+                android:background="@color/transparent_grey"
+                android:textColorHint="@color/dark_grey"
+                android:hint="@string/hint_search"
+                android:paddingStart="@dimen/grid_2x"
+                android:paddingEnd="@dimen/grid_2x"
+                android:textColor="@color/black"
+                android:drawableEnd="@android:drawable/ic_menu_search"
+                android:imeOptions="actionSearch"
+                android:inputType="text"/>
+
+        </androidx.cardview.widget.CardView>
+
+    </androidx.cardview.widget.CardView>
 
     <Button
         android:id="@+id/logoutButton"
+        style="@style/StandardButton"
         android:layout_width="@dimen/standard_width"
         android:layout_height="@dimen/standard_height"
         android:onClick="onLogout"
         android:text="@string/button_logout"
-        style="@style/StandardButton"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
