+++ app/src/main/res/values/dimens.xml
@@ -9,4 +9,10 @@
     <dimen name="standard_height">50dp</dimen>
 
     <dimen name="standard_text_size">20sp</dimen>
+
+    <dimen name="standard_corner_radius">25dp</dimen>
+
+    <dimen name="standard_image">50dp</dimen>
+
+    <dimen name="search_height">40dp</dimen>
 </resources>
