+++ app/src/main/res/values/strings.xml
@@ -9,4 +9,5 @@
     <string name="button_logout">Logout</string>
     <string name="button_signup">Signup</string>
     <string name="text_login">Already have an account? Go to Login</string>
+    <string name="hint_search">Search</string>
 </resources>
