+++ app/src/main/res/values/themes.xml
@@ -27,4 +27,10 @@
         <item name="android:textColor">@color/blue</item>
     </style>
 
+    <style name="Heading">
+        <item name="android:textSize">26sp</item>
+        <item name="android:textStyle">bold</item>
+        <item name="android:textColor">@color/black</item>
+    </style>
+
 </resources>
